+++ Fabric/src/main/java/fuzs/resourcepackoverrides/client/ResourcePackOverridesFabricClient.java
@@ -1,6 +1,6 @@
 package fuzs.resourcepackoverrides.client;
 
-import fuzs.resourcepackoverrides.client.handler.PackIdTooltipHandler;
+import fuzs.resourcepackoverrides.client.handler.PackActionsHandler;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.screen.v1.ScreenEvents;
@@ -17,10 +17,10 @@ public class ResourcePackOverridesFabricClient implements ClientModInitializer {
     private static void registerHandlers() {
         ScreenEvents.AFTER_INIT.register((client, screen, scaledWidth, scaledHeight) -> {
             if (screen instanceof PackSelectionScreen) {
-                ScreenEvents.afterRender(screen).register(PackIdTooltipHandler::onScreen$Render$Post);
-                ScreenKeyboardEvents.afterKeyPress(screen).register(PackIdTooltipHandler::onKeyPressed$Post);
+                ScreenEvents.afterRender(screen).register(PackActionsHandler::onScreen$Render$Post);
+                ScreenKeyboardEvents.afterKeyPress(screen).register(PackActionsHandler::onKeyPressed$Post);
             }
         });
-        ClientTickEvents.END_CLIENT_TICK.register(PackIdTooltipHandler::onClientTick$End);
+        ClientTickEvents.END_CLIENT_TICK.register(PackActionsHandler::onClientTick$End);
     }
 }