+++ Forge/src/main/java/fuzs/resourcepackoverrides/client/ResourcePackOverridesForgeClient.java
@@ -1,11 +1,13 @@
 package fuzs.resourcepackoverrides.client;
 
 import fuzs.resourcepackoverrides.ResourcePackOverrides;
-import fuzs.resourcepackoverrides.client.handler.PackIdTooltipHandler;
+import fuzs.resourcepackoverrides.client.handler.PackActionsHandler;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.packs.PackSelectionScreen;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.ScreenEvent;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLConstructModEvent;
@@ -21,13 +23,16 @@ public class ResourcePackOverridesForgeClient {
     private static void registerHandlers() {
         MinecraftForge.EVENT_BUS.addListener((final ScreenEvent.Render.Post evt) -> {
             if (evt.getScreen() instanceof PackSelectionScreen) {
-                PackIdTooltipHandler.onScreen$Render$Post(evt.getScreen(), evt.getPoseStack(), evt.getMouseX(), evt.getMouseY(), evt.getPartialTick());
+                PackActionsHandler.onScreen$Render$Post(evt.getScreen(), evt.getPoseStack(), evt.getMouseX(), evt.getMouseY(), evt.getPartialTick());
             }
         });
         MinecraftForge.EVENT_BUS.addListener((final ScreenEvent.KeyPressed.Post evt) -> {
             if (evt.getScreen() instanceof PackSelectionScreen) {
-                PackIdTooltipHandler.onKeyPressed$Post(evt.getScreen(), evt.getKeyCode(), evt.getScanCode(), evt.getModifiers());
+                PackActionsHandler.onKeyPressed$Post(evt.getScreen(), evt.getKeyCode(), evt.getScanCode(), evt.getModifiers());
             }
+        });
+        MinecraftForge.EVENT_BUS.addListener((final TickEvent.ClientTickEvent evt) -> {
+            if (evt.phase == TickEvent.Phase.END) PackActionsHandler.onClientTick$End(Minecraft.getInstance());
         });
     }
 }